Finalizes DB test setup
