+++ src/golang/lib/repos/tests/seed.go
@@ -1,0 +1,10 @@
+package tests
+
+import "github.com/aqueducthq/aqueduct/lib/models"
+
+// seedWorkflow creates count workflow records.
+// It also creates the necessary user records.
+func (ts *TestSuite) seedWorkflow(count int) []models.Workflow {
+	// TODO: After User refactor is complete
+	return nil
+}
