+++ src/golang/lib/repos/tests/workflow.go
@@ -1,0 +1,7 @@
+package tests
+
+func (ts *TestSuite) TestWorkflow_Exists() {
+	// TODO: Fix test once user refactor is complete
+	workflows := ts.seedWorkflow(1)
+	ts.Require().Nil(workflows)
+}
